feat: Refactor deleteContactForm action and file structure
The deleteContactForm action has been refactored to improve code organization and maintainability. The action has been moved to a new file named "delete.js" in the "dashboard-actions" folder. This change helps to keep related actions together and makes it easier to locate and modify the code in the future.
