+++ actions/dashboard-actions/delete.js
